Delta versioon 3.3.1.22
Koos Delta versiooni paigaldamisega on vajalik paigaldada ka mso-service versioon.


git-svn-id: https://svn.smit/tarkvara/edhs/trunk@108663 5fddf92a-9921-11dd-8d32-b999f448c456

+++ delta/common/source/java/ee/webmedia/alfresco/classificator/constant/FieldType.java
@@ -21,7 +21,7 @@ public enum FieldType {
     /** Valikuväli */
     COMBOBOX(Props.CLASSIFICATOR, Props.CLASSIFICATOR_DEFAULT_VALUE),
     /** Täisarv */
-    INT(Props.DEFAULT_VALUE),
+    LONG(Props.DEFAULT_VALUE),
     /** Komakohaga arv */
     DOUBLE(Props.DEFAULT_VALUE),
     /** Kuupäev */
